+++ lib/src/async_value/async_value.dart
@@ -1,2 +1,3 @@
 export 'async_value_cubit.dart';
 export 'async_value_wrapper.dart';
+export 'hydrated_async_value_cubit.dart';