+++ lib/src/async_value/hydrated_async_value_cubit.dart
@@ -3,9 +3,9 @@ import 'package:hydrated_bloc/hydrated_bloc.dart';
 
 /// Base class for Cubit that emits a [AsyncValueWrapper]
 /// This is made to make emit of the value easier in subclasses
-abstract class AsyncValueCubit<T> extends HydratedCubit<AsyncValueWrapper<T>> {
+abstract class HydratedAsyncValueCubit<T> extends HydratedCubit<AsyncValueWrapper<T>> {
   /// Constructor
-  AsyncValueCubit() : super(AsyncValueWrapper<T>.initial());
+  HydratedAsyncValueCubit() : super(AsyncValueWrapper<T>.initial());
 
   /// Emits a [AsyncValueWrapper.loading] value
   /// if there was a value before, it will be passed as oldValue
@@ -51,5 +51,25 @@ abstract class AsyncValueCubit<T> extends HydratedCubit<AsyncValueWrapper<T>> {
     fn()
         .then((value) => emitSuccess(value))
         .catchError((err) => emitError(err));
+  }
+
+  /// Helper function that is useful to parse the value of the [AsyncValueWrapper] state
+  /// In most cases, if your entity/model has a factory method you can use
+  /// `Model.fromJson`
+  abstract T Function(dynamic json) valueFromJson;
+
+  /// Helper function that is useful to serialize the value of the [AsyncValueWrapper] state
+  /// In most cases, if your entity/model has a toJson method you can use
+  /// `t.toJson`
+  abstract dynamic Function(T t) valueToJson;
+
+  @override
+  AsyncValueWrapper<T>? fromJson(Map<String, dynamic> json) {
+    return AsyncValueWrapper.fromJson(json, valueFromJson);
+  }
+
+  @override
+  Map<String, dynamic>? toJson(AsyncValueWrapper<T> state) {
+    return state.toJson(valueToJson);
   }
 }