+++ CHANGELOG.md
@@ -1,4 +1,17 @@
-## 1.2.0
+## 2.0.0
+
+### Added
+
+- Added dedicated identifier for `AsyncValueStatus`
+- Added from json factory for `AsyncValueStatus`
+- Added from json factory for `AsyncValueWrapper` using generic factory
+- Added to json  `AsyncValueWrapper` using parsing method
+- Added `HydratedAsyncValueCubit`
+
+### Breaking change
+
+- Added automatic serialisation / parsing in `HydratedAsyncValueCubit`
+- Required `valueFromJson` and `valueToJson` implementation in `HydratedAsyncValueCubit`
 
 ## 1.1.0